refactor: isolate feature provenance
