+++ plugin/identity.json
@@ -2,5 +2,5 @@
   "schema_version": 1,
   "plugin_version": "0.74.5",
   "hook_manifest_sha256": "ea53d65130d73a4a438d8e9040a17462db52fa3c23e2a0331ed87268543644f4",
-  "inventory_sha256": "480cb5d940b32f90e3873f50bdb86fbae889290fb42a718ea85d1bccd3e95bbc"
+  "inventory_sha256": "91c4d28e96614635504e3d7d6af4532cb11d8c57140cad9012be03f27ea0453e"
 }
